requirements.txt
plotly==4.7.0
matplotlib==3.0.3
scikit_learn==0.22.2.post1
scipy==1.4.1
wfdb==2.2.1
albumentations==0.4.3
efficientnet_pytorch==0.6.3
torchvision==0.3.0
opencv_python==4.2.0.32

